Don't use message factory when resetting passwords.
Use the object initialiser on MailMessage instead of creating one via the mostly-redundant message factory.
This required changing the ResetPasswordEmailSender tests to test that a specific message is sent, rather
than testing that a specific message is created (and possibly never sent). Additionally, we had to add
equality to MailMessage so that NSubstitute could do its stuff... We used R#'s equality implementation for
the sake of speed (with apologies to @adrianbanks).