Extract IMessenger and IMailTemplates.
IMessenger notifies a user about particular events on the site that they are interested in.
For example, SessionCreationMailMessenger sends a user the abstract of the talk they have just
submitted.

IMailTemplate can render a given file to a string for transmission by an IMessenger. The
TokenSubstitutingMailTemplate reads the file and replaces the specified tokens with the specified
substitutions; the MarkdownConvertingMailTemplate renders the output in Markdown.